requirements.txt
annotated-types==0.7.0
anyio==4.4.0
backoff==2.2.1
certifi==2024.7.4
charset-normalizer==3.3.2
distro==1.9.0
exceptiongroup==1.2.2
h11==0.14.0
haystack-ai==2.3.1
haystack-experimental==0.1.1
httpcore==1.0.5
httpx==0.27.0
idna==3.7
Jinja2==3.1.4
jiter==0.5.0
lazy-imports==0.3.1
MarkupSafe==2.1.5
monotonic==1.6
more-itertools==10.3.0
networkx==3.3
numpy==1.26.4
openai==1.40.0
pandas==2.2.2
posthog==3.5.0
pydantic==2.8.2
pydantic_core==2.20.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
six==1.16.0
sniffio==1.3.1
tenacity==9.0.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
